+++ src/views/DaoInfo/components/DAOHoldingsDetails/DAOHoldingsDetails.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useCallback, useEffect } from "react";
 import Chip from "@mui/material/Chip";
-import Grid from "@mui/material/Grid";
 import styled from "styled-components";
 import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
@@ -44,52 +43,34 @@ const DAOHoldingsDetails: React.FC = () => {
     <>
       <div style={{ clear: "both", margin: "10px" }}></div>
 
-      <Typography variant="h5">DAO Holdings</Typography>
-
+      <Typography variant="h5">
+        DAO Holdings <InfoIconWithTooltip tooltipText="The DAO holds these coins and tokens" />
+      </Typography>
+      <br />
       <StyledBox>
-        <Typography align="center" variant="h6">
-          Coins/Tokens <InfoIconWithTooltip tooltipText="The DAO holds these coins and tokens" />
-        </Typography>
-        <br />
-        <Grid container sx={{ textAlign: "center" }} spacing={1}>
-          <Grid xs={4} item>
-            <Chip label={"UBQ: " + daoHoldings.ubq.toFormat(0)} color="primary" variant="outlined" />
-          </Grid>
-          <Grid xs={4} item>
-            <Chip label={"INK: " + daoHoldings.ink.toFormat(0)} color="primary" variant="outlined" />
-          </Grid>
-          <Grid xs={4} item>
-            <Chip label={"ESCH: " + daoHoldings.esch.toFormat(0)} color="primary" variant="outlined" />
-          </Grid>
-        </Grid>
-
-        <Typography sx={{ mt: "12px" }} align="center" variant="h6">
-          LP Standing <InfoIconWithTooltip tooltipText="The DAO holds this LP" />
-        </Typography>
-        <br />
-        <Grid container sx={{ textAlign: "center", clear: "both" }} spacing={1}>
-          <Grid xs={4} item>
-            <Chip label={"UBQ/INK: " + daoHoldings.lp.ubqInk.toFormat(4)} color="primary" variant="outlined" />
-          </Grid>
-          <Grid xs={4} item>
-            <Chip label={"GRANS/INK: " + daoHoldings.lp.gransInk.toFormat(4)} color="primary" variant="outlined" />
-          </Grid>
-          <Grid xs={4} item>
-            <Chip label={"INK/ESCH: " + daoHoldings.lp.inkEsch.toFormat(4)} color="primary" variant="outlined" />
-          </Grid>
-        </Grid>
+        <Chip label={"UBQ: " + daoHoldings.ubq.toFormat(0)} color="primary" variant="outlined" />
+        <Chip label={"INK: " + daoHoldings.ink.toFormat(0)} color="primary" variant="outlined" />
+        <Chip label={"ESCH: " + daoHoldings.esch.toFormat(0)} color="primary" variant="outlined" />
       </StyledBox>
     </>
   );
 };
 
 const StyledBox = styled(Box)`
-  display: block;
+  display: flex;
+  justify-content: center;
+  gap: 8%;
   border-radius: 15px;
   background: #2c2b3d;
   padding: 20px;
-  width: 60%;
-  min-width: 550px;
+  width: 50%;
+  position: relative;
+  min-width: 650px;
+
+  @media (max-width: 850px) {
+    flex-direction: column;
+    min-width: 600px;
+  }
 `;
 
 export default React.memo(DAOHoldingsDetails);